refactor: domain 폴더로 구조 변경
