+++ src/main/kotlin/lotto/domain/BonusLotto.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class BonusLotto(private val bonusNumber: LottoNumber) {
