+++ src/main/kotlin/lotto/domain/LottoPurchaseAmount.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class LottoPurchaseAmount(private val lottoPurchaseAmount:Any) {
 //    init {
